refactor: check in dedicated classes
